Rename service folder to lib, add bold style for email text - #88
