+++ src/components/AppLogin/index.js
@@ -1,12 +1,7 @@
 import React, { useContext } from 'react'
-import { Roboto } from 'next/font/google'
 import AppContext from '../../context/AppContext'
-import {MailOutlined} from '@ant-design/icons'
-
-const roboto_light = Roboto( {
-    weight: '300',
-    subsets: ['latin']
-})
+import {MailOutlined, UserOutlined, LoginOutlined} from '@ant-design/icons'
+import {robotoLight, robotoBold } from "../../lib/fonts";
 
 const AppLogin = ({ onLogin, unauthorized, onLogout }) => {
 
@@ -15,12 +10,12 @@ const AppLogin = ({ onLogin, unauthorized, onLogout }) => {
     const pageData = () => {
         if (unauthorized) {
             const body = <div>
-                <p>You are trying to access the {t('HuBMAP')} Data Ingest Board logged in as <strong className='font-weight-bold'>{getUserEmail()}</strong>.
+                <p>You are trying to access the {t('HuBMAP')} Data Ingest Board logged in as <strong className={robotoBold.className}>{getUserEmail()}</strong>.
                 You are not authorized to log into this application with that account.  Please check that
-                you have registered via the <a href={t('https://profile.hubmapconsortium.org/profile')}>Member Registration Page</a>,
+                you have registered via the <a href={t('https://profile.hubmapconsortium.org/profile')} className='lnk--ic'>Member Registration Page <UserOutlined /></a>,
                 have provided Globus account information and are logging in with that account.</p>
 
-                <p>Once you have confirmed your registration information you can <a href='/login'>log in</a> again.</p>
+                <p>Once you have confirmed your registration information you can <a href='/login' className='lnk--ic'>log in <LoginOutlined /></a> again.</p>
 
                 <p>If you continue to have issues accessing this site please contact the &nbsp;
                     <a href={`mailto:${t('help@hubmapconsortium.org')}`} className='lnk--ic'>{t('HuBMAP')} Help Desk <MailOutlined /></a>.
@@ -47,7 +42,7 @@ const AppLogin = ({ onLogin, unauthorized, onLogout }) => {
             <div className="container">
                 <div className={`c-login row mt-4 ${unauthorized ? 'alert alert-danger' : ''}`}>
                     <h1 className="c-login__head col-6">{t(details.title, [t('HuBMAP')])}</h1>
-                    <div className={`c-login__txt ${roboto_light.className}`}>
+                    <div className={`c-login__txt ${robotoLight.className}`}>
                         {details.body}
                     </div>
                     <div className="row mt-3">
