+++ src/lib/fonts.js
@@ -1,0 +1,14 @@
+import {Roboto} from "next/font/google";
+
+
+const robotoLight = Roboto( {
+    weight: '300',
+    subsets: ['latin']
+})
+
+const robotoBold = Roboto( {
+    weight: '900',
+    subsets: ['latin']
+})
+
+export {robotoLight, robotoBold }
